sla-fw/requirements.txt
evdev
readerwriterlock
distro
bitstring
gpio
numpy
paho-mqtt
pydbus
pygobject
pyserial
toml
PySignal
Pillow
psutil
setuptools
deprecation
requests
aiohttp
pywayland
smbus2
gpiod==1.5.4
pyyaml
git+https://github.com/prusa3d/Prusa-Error-Codes@master#egg=prusa-errors
